Add files via upload
As Felorence dataset includes video streams with different length, firstly I tried to preprocess the streams to existing maximum length( this is optional and you can try the average or minimum length) then in 2 stages I implemented a model using just LSTM network and a fully-connected dense. Results are displayed well and easy to compare.